Don't automatically create the config on launch
This fixes something that has annoyed me for a while. If you removed
your config file (specifically a symlink) while the app was running, we
would automatically create an empty one. This changes it to only create
a new config file on launch. If you remove yours while running you will
get a warning explaining the file is missing and relaunch to create one.